requirements/base.txt
# Core dependencies
numpy>=1.21.0
click>=8.0.0           # CLI
Pillow>=9.0.0          # Image processing
pypdf>=3.0.0           # PDF parsing
zstandard>=0.19.0      # ZSTD compression
brotli>=1.0.9          # Brotli compression

# Data structures
msgpack>=1.0.0         # Efficient serialization
xxhash>=3.0.0          # Fast hashing
